+++ D3DGame_181205_084_1_PostEffect/Executes/TestCompute.cpp
@@ -12,11 +12,11 @@ TestCompute::TestCompute()
 
 	for (int i = 0; i < dataSize; i++)
 	{
-		dataA[i].V1 = D3DXVECTOR3(i, i, i);
-		dataA[i].V2 = D3DXVECTOR2(i, 0.0f);
+		dataA[i].V1 = D3DXVECTOR3((float)i, (float)i, (float)i);
+		dataA[i].V2 = D3DXVECTOR2((float)i, 0.0f);
 
-		dataB[i].V1 = D3DXVECTOR3(-i, i, i);
-		dataB[i].V2 = D3DXVECTOR2(i, -i);
+		dataB[i].V1 = D3DXVECTOR3((float)(-i), (float)i, (float)i);
+		dataB[i].V2 = D3DXVECTOR2((float)i, (float)(-i));
 	}
 
 	input[0] = new CsResource(sizeof(Data), dataSize, &dataA[0]);
@@ -67,7 +67,7 @@ void TestCompute::ExecuteCs()
 	vector<Data> dataView(dataSize);
 	output->Read(&dataView[0]);
 
-	for (int i = 0; i < dataSize; i++)
+	for (UINT i = 0; i < dataSize; i++)
 	{
 		fprintf(file, "V1 : (%f, %f, %f)", dataView[i].V1.x, dataView[i].V1.y, dataView[i].V1.z);
 		fprintf(file, " - V2 : (%f, %f)\n", dataView[i].V2.x, dataView[i].V2.y);
